requirements.txt
backoff==2.2.1
black==24.2.0
boto3==1.34.52
openpyxl==3.1.2
pandas==1.5.3
numpy==1.26.4
pydantic>2.0.0
pyyaml==6.0.1
requests==2.31.0
xlsxwriter==3.2.0
google-auth 
google-auth-oauthlib 
google-api-python-client
# Testing
pytest
pytest-env
pytest-mock